Python/requirements.txt
# -----------------------------------------------------------------
# Python Dependencies for the Pepper VR Teleoperation Server
# Updated: October 2025
# To install: pip install -r requirements.txt
# -----------------------------------------------------------------

# Core library for Pepper robot communication. Version is fixed by manufacturer.
naoqi-python==2.8.7.4

# WebSocket library for real-time commands.
websockets==13.0

# Web framework for streaming the camera feed.
Flask==3.1.2

# Flask extension for handling Cross-Origin Resource Sharing.
Flask-Cors==6.0.1

# Fundamental library for numerical operations (image data).
# Pinned to 1.26.x as it's a very stable Long-Term Support-like branch.
numpy==1.26.4

# Computer vision library for processing camera frames.
opencv-python==4.12.0.88